Extract EDP simulator for EventGroups with legacy metadata
