+++ .github/workflows/configure-simulators.yml
@@ -138,14 +138,14 @@ jobs:
     - name: Generate archives
       run: >
         bazelisk build 
-        //src/main/k8s/dev:synthetic_generator_edp_simulators.tar
+        //src/main/k8s/dev:edp_simulators.tar
         //src/main/k8s/testing/secretfiles:kustomization.tar
 
     - name: Make Kustomization dir
       run: mkdir -p "$KUSTOMIZATION_PATH"
     - name: Extract Kustomization archive
       run: >
-        tar -xf "$BAZEL_BIN/src/main/k8s/dev/synthetic_generator_edp_simulators.tar"
+        tar -xf "$BAZEL_BIN/src/main/k8s/dev/edp_simulators.tar"
         -C "$KUSTOMIZATION_PATH"
 
     - name: Extract secret files archive
@@ -154,7 +154,7 @@ jobs:
         -C "$KUSTOMIZATION_PATH"
 
     - name: Export KUSTOMIZE_PATH
-      run: echo "KUSTOMIZE_PATH=$KUSTOMIZATION_PATH/src/main/k8s/dev/synthetic_generator_edp_simulators" >> $GITHUB_ENV
+      run: echo "KUSTOMIZE_PATH=$KUSTOMIZATION_PATH/src/main/k8s/dev/edp_simulators" >> $GITHUB_ENV
 
     # Run kubectl diff, treating the command as succeeded even if the exit
     # code is 1 as kubectl uses this code to indicate there's a diff.
