+++ src/main/k8s/dev/BUILD.bazel
@@ -347,16 +347,6 @@ kustomization_dir(
     ],
 )
 
-cue_library(
-    name = "edp_simulator_gke",
-    srcs = ["edp_simulator_gke.cue"],
-    deps = [
-        ":base_gke",
-        ":config_gke",
-        "//src/main/k8s:edp_simulator",
-    ],
-)
-
 EDP_SIMULATOR_TAGS = {
     "secret_name": SECRET_NAME,
     "mc_name": SIMULATOR_K8S_SETTINGS.mc_name,
@@ -384,37 +374,41 @@ EDP_SIMULATOR_TAGS = {
 }
 
 cue_dump(
-    name = "synthetic_generator_edp_simulator_gke",
-    srcs = ["synthetic_generator_edp_simulator_gke.cue"],
+    name = "edp_simulators_gke",
+    srcs = ["edp_simulators_gke.cue"],
     cue_tags = EDP_SIMULATOR_TAGS,
     tags = ["manual"],
-    deps = [":edp_simulator_gke"],
+    deps = [
+        ":base_gke",
+        ":config_gke",
+        "//src/main/k8s:edp_simulator",
+    ],
 )
 
 kustomization_dir(
-    name = "synthetic_generator_config_files",
+    name = "edp_simulator_config_files",
     testonly = True,
     srcs = [
-        "synthetic_generator_config_files_kustomization.yaml",
+        "edp_simulator_config_files.yaml",
         "//src/main/k8s/testing/data:synthetic_generation_specs_large",
         "//src/main/k8s/testing/secretfiles:known_event_group_metadata_type_set",
     ],
     renames = {
-        "synthetic_generator_config_files_kustomization.yaml": "kustomization.yaml",
+        "edp_simulator_config_files.yaml": "kustomization.yaml",
     },
 )
 
 kustomization_dir(
-    name = "synthetic_generator_edp_simulators",
+    name = "edp_simulators",
     testonly = True,
     srcs = [
         "resource_requirements.yaml",
-        ":synthetic_generator_edp_simulator_gke",
+        ":edp_simulators_gke",
     ],
     generate_kustomization = True,
     tags = ["manual"],
     deps = [
-        ":synthetic_generator_config_files",
+        ":edp_simulator_config_files",
         "//src/main/k8s/testing/secretfiles:kustomization",
     ],
 )
