+++ src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider/EdpSimulatorFlags.kt
@@ -16,6 +16,7 @@ package org.wfanet.measurement.loadtest.dataprovider
 
 import java.io.File
 import java.time.Duration
+import java.time.ZoneId
 import org.wfanet.measurement.common.grpc.TlsFlags
 import org.wfanet.measurement.loadtest.KingdomPublicApiFlags
 import org.wfanet.measurement.loadtest.RequisitionFulfillmentServiceFlags
@@ -105,6 +106,47 @@ class EdpSimulatorFlags {
     private set
 
   @CommandLine.Option(
+    names = ["--event-message-descriptor-set"],
+    description =
+      [
+        "Serialized FileDescriptorSet for the event message and its dependencies.",
+        "This can be specified multiple times.",
+        "It need not be specified if the event message type is $TEST_EVENT_MESSAGE_TYPE.",
+      ],
+    required = false,
+  )
+  var eventMessageDescriptorSetFiles: List<File> = emptyList()
+    private set
+
+  @CommandLine.Option(
+    names = ["--population-spec"],
+    description = ["Path to SyntheticPopulationSpec message in text format."],
+    required = true,
+  )
+  lateinit var populationSpecFile: File
+    private set
+
+  lateinit var syntheticDataTimeZone: ZoneId
+    private set
+
+  @CommandLine.Option(
+    names = ["--synthetic-data-time-zone"],
+    description = ["ID of time zone for synthetic data generation"],
+    required = false,
+    defaultValue = "UTC",
+  )
+  private fun setSyntheticDataTimeZone(zoneId: String) {
+    syntheticDataTimeZone = ZoneId.of(zoneId)
+  }
+
+  @CommandLine.Option(
+    names = ["--support-hmss"],
+    description = ["Whether to support HMSS protocol requisitions."],
+  )
+  var supportHmss: Boolean = false
+    private set
+
+  @CommandLine.Option(
     names = ["--random-seed"],
     description = ["Random seed of differential privacy noisers for direct measurements"],
     required = false,
@@ -131,4 +173,9 @@ class EdpSimulatorFlags {
   )
   var healthFile: File? = null
     private set
+
+  companion object {
+    const val TEST_EVENT_MESSAGE_TYPE =
+      "wfa.measurement.api.v2alpha.event_templates.testing.TestEvent"
+  }
 }
