+++ src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider/SyntheticGeneratorEventQuery.kt
@@ -19,6 +19,7 @@ package org.wfanet.measurement.loadtest.dataprovider
 import com.google.protobuf.Descriptors
 import com.google.protobuf.DynamicMessage
 import com.google.protobuf.TypeRegistry
+import java.time.ZoneId
 import org.projectnessie.cel.Program
 import org.wfanet.measurement.api.v2alpha.EventGroup
 import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.SyntheticEventGroupSpec
@@ -31,6 +32,7 @@ import org.wfanet.measurement.eventdataprovider.eventfiltration.EventFilters
 abstract class SyntheticGeneratorEventQuery(
   val populationSpec: SyntheticPopulationSpec,
   private val eventMessageDescriptor: Descriptors.Descriptor,
+  val timeZone: ZoneId,
 ) : EventQuery<DynamicMessage> {
   init {
     require(
@@ -43,7 +45,12 @@ abstract class SyntheticGeneratorEventQuery(
   constructor(
     populationSpec: SyntheticPopulationSpec,
     typeRegistry: TypeRegistry,
-  ) : this(populationSpec, typeRegistry.getDescriptorForTypeUrl(populationSpec.eventMessageTypeUrl))
+    timeZone: ZoneId,
+  ) : this(
+    populationSpec,
+    typeRegistry.getDescriptorForTypeUrl(populationSpec.eventMessageTypeUrl),
+    timeZone,
+  )
 
   /** Returns the synthetic data spec for [eventGroup]. */
   protected abstract fun getSyntheticDataSpec(eventGroup: EventGroup): SyntheticEventGroupSpec
@@ -60,6 +67,7 @@ abstract class SyntheticGeneratorEventQuery(
         populationSpec,
         syntheticDataSpec,
         timeRange,
+        timeZone,
       )
       .flatMap { it.labeledEvents }
       .filter { EventFilters.matches(it.message, program) }
