+++ src/test/kotlin/org/wfanet/measurement/loadtest/dataprovider/BUILD.bazel
@@ -4,8 +4,8 @@ load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library", "kt_jvm_test")
 package(default_testonly = True)
 
 kt_jvm_library(
-    name = "edp_simulator_test",
-    srcs = ["EdpSimulatorTest.kt"],
+    name = "abstract_edp_simulator_test",
+    srcs = ["AbstractEdpSimulatorTest.kt"],
     data = [
         "//src/main/k8s/testing/secretfiles:all_configs",
         "//src/main/k8s/testing/secretfiles:all_der_files",
@@ -13,6 +13,35 @@ kt_jvm_library(
         "//src/main/k8s/testing/secretfiles:edp_trusted_certs.pem",
     ],
     deps = [
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement:privacy_budget_manager",
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/testing",
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/requisition/v2alpha:requisition",
+        "//src/main/kotlin/org/wfanet/measurement/integration/common:synthetic_generation_specs",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/config:test_identifiers",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:abstract_edp_simulator",
+        "//src/main/proto/wfa/measurement/api/v2alpha:api_kt_jvm_grpc_proto",
+        "@wfa_common_jvm//imports/java/com/google/protobuf",
+        "@wfa_common_jvm//imports/java/org/junit",
+        "@wfa_common_jvm//imports/kotlin/com/google/protobuf/kotlin",
+        "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
+        "@wfa_common_jvm//imports/kotlin/org/mockito/kotlin",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:hashing",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:security_provider",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto:signing_key_storage",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/testing",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc/testing",
+        "@wfa_consent_signaling_client//src/main/kotlin/org/wfanet/measurement/consent/client/duchy",
+        "@wfa_consent_signaling_client//src/main/kotlin/org/wfanet/measurement/consent/client/measurementconsumer",
+    ],
+)
+
+kt_jvm_library(
+    name = "edp_simulator_test",
+    srcs = ["EdpSimulatorTest.kt"],
+    deps = [
+        ":abstract_edp_simulator_test",
         "//src/main/kotlin/org/wfanet/measurement/api/v2alpha/testing",
         "//src/main/kotlin/org/wfanet/measurement/dataprovider:measurement_results",
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/eventfiltration:event_filters",
@@ -46,6 +75,38 @@ java_test(
     name = "EdpSimulatorTest",
     test_class = "org.wfanet.measurement.loadtest.dataprovider.EdpSimulatorTest",
     runtime_deps = [":edp_simulator_test"],
+)
+
+kt_jvm_library(
+    name = "legacy_metadata_edp_simulator_test",
+    srcs = ["LegacyMetadataEdpSimulatorTest.kt"],
+    deps = [
+        ":abstract_edp_simulator_test",
+        "//src/main/kotlin/org/wfanet/measurement/integration/common:synthetic_generation_specs",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/config:event_group_metadata",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/config:test_identifiers",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:in_memory_event_query",
+        "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:legacy_metadata_edp_simulator",
+        "//src/main/proto/wfa/measurement/api/v2alpha:api_kt_jvm_grpc_proto",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
+        "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
+        "@wfa_common_jvm//imports/java/org/junit",
+        "@wfa_common_jvm//imports/kotlin/kotlin/test",
+        "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
+        "@wfa_common_jvm//imports/kotlin/org/mockito/kotlin",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc/testing",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/testing",
+        "@wfa_consent_signaling_client//src/main/kotlin/org/wfanet/measurement/consent/client/measurementconsumer",
+    ],
+)
+
+# TODO(bazelbuild/rules_kotlin#1088): Use kt_jvm_test when fixed.
+# Using kt_jvm_test directly does not correctly set the the path for loading native libraries.
+java_test(
+    name = "LegacyMetadataEdpSimulatorTest",
+    test_class = "org.wfanet.measurement.loadtest.dataprovider.LegacyMetadataEdpSimulatorTest",
+    runtime_deps = [":legacy_metadata_edp_simulator_test"],
 )
 
 kt_jvm_test(
