+++ src/test/kotlin/org/wfanet/measurement/loadtest/dataprovider/EdpSimulatorTest.kt
@@ -18,76 +18,46 @@ import com.google.common.truth.Truth.assertThat
 import com.google.common.truth.extensions.proto.FieldScopes
 import com.google.common.truth.extensions.proto.ProtoTruth.assertThat
 import com.google.protobuf.ByteString
-import com.google.protobuf.kotlin.toByteString
 import com.google.protobuf.kotlin.toByteStringUtf8
+import com.google.protobuf.timestamp
 import com.google.type.interval
 import io.grpc.Status
-import java.nio.file.Path
-import java.nio.file.Paths
-import java.security.cert.X509Certificate
 import java.time.Clock
 import java.time.Duration
 import java.time.LocalDate
 import java.time.ZoneOffset
 import kotlin.random.Random
-import kotlin.test.assertFailsWith
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
-import org.junit.Rule
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
 import org.mockito.kotlin.any
-import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
 import org.mockito.kotlin.stub
-import org.mockito.kotlin.times
 import org.mockito.kotlin.verifyBlocking
 import org.wfanet.anysketch.Sketch
 import org.wfanet.anysketch.crypto.ElGamalPublicKey
 import org.wfanet.frequencycount.FrequencyVector
-import org.wfanet.measurement.api.v2alpha.CertificatesGrpcKt.CertificatesCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.CertificatesGrpcKt.CertificatesCoroutineStub
-import org.wfanet.measurement.api.v2alpha.CreateEventGroupMetadataDescriptorRequest
 import org.wfanet.measurement.api.v2alpha.CreateEventGroupRequest
-import org.wfanet.measurement.api.v2alpha.DataProviderCertificateKey
-import org.wfanet.measurement.api.v2alpha.DataProvidersGrpcKt.DataProvidersCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.DataProvidersGrpcKt.DataProvidersCoroutineStub
-import org.wfanet.measurement.api.v2alpha.DuchyCertificateKey
-import org.wfanet.measurement.api.v2alpha.DuchyKey
 import org.wfanet.measurement.api.v2alpha.EventGroup
 import org.wfanet.measurement.api.v2alpha.EventGroupKey
-import org.wfanet.measurement.api.v2alpha.EventGroupMetadataDescriptorsGrpcKt.EventGroupMetadataDescriptorsCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.EventGroupMetadataDescriptorsGrpcKt.EventGroupMetadataDescriptorsCoroutineStub
+import org.wfanet.measurement.api.v2alpha.EventGroupMetadata
+import org.wfanet.measurement.api.v2alpha.EventGroupMetadataKt
 import org.wfanet.measurement.api.v2alpha.EventGroupsGrpcKt.EventGroupsCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.EventGroupsGrpcKt.EventGroupsCoroutineStub
 import org.wfanet.measurement.api.v2alpha.FulfillDirectRequisitionRequest
 import org.wfanet.measurement.api.v2alpha.FulfillRequisitionRequest
 import org.wfanet.measurement.api.v2alpha.FulfillRequisitionRequestKt
-import org.wfanet.measurement.api.v2alpha.FulfillRequisitionResponse
-import org.wfanet.measurement.api.v2alpha.GetEventGroupRequest
 import org.wfanet.measurement.api.v2alpha.Measurement
-import org.wfanet.measurement.api.v2alpha.MeasurementConsumersGrpcKt
-import org.wfanet.measurement.api.v2alpha.MeasurementConsumersGrpcKt.MeasurementConsumersCoroutineStub
 import org.wfanet.measurement.api.v2alpha.MeasurementSpec
 import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.impression
 import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.reach
-import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.reachAndFrequency
 import org.wfanet.measurement.api.v2alpha.MeasurementSpecKt.vidSamplingInterval
-import org.wfanet.measurement.api.v2alpha.PopulationSpecKt.subPopulation
-import org.wfanet.measurement.api.v2alpha.PopulationSpecKt.vidRange
+import org.wfanet.measurement.api.v2alpha.MediaType
 import org.wfanet.measurement.api.v2alpha.ProtocolConfig
 import org.wfanet.measurement.api.v2alpha.ProtocolConfigKt
 import org.wfanet.measurement.api.v2alpha.RefuseRequisitionRequest
-import org.wfanet.measurement.api.v2alpha.ReplaceDataAvailabilityIntervalRequest
 import org.wfanet.measurement.api.v2alpha.Requisition
 import org.wfanet.measurement.api.v2alpha.Requisition.Refusal
-import org.wfanet.measurement.api.v2alpha.RequisitionFulfillmentGrpcKt.RequisitionFulfillmentCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.RequisitionFulfillmentGrpcKt.RequisitionFulfillmentCoroutineStub
 import org.wfanet.measurement.api.v2alpha.RequisitionKt.DuchyEntryKt.honestMajorityShareShuffle
-import org.wfanet.measurement.api.v2alpha.RequisitionKt.DuchyEntryKt.liquidLegionsV2
 import org.wfanet.measurement.api.v2alpha.RequisitionKt.DuchyEntryKt.value
 import org.wfanet.measurement.api.v2alpha.RequisitionKt.duchyEntry
 import org.wfanet.measurement.api.v2alpha.RequisitionKt.refusal
@@ -96,128 +66,51 @@ import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt
 import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt.eventFilter
 import org.wfanet.measurement.api.v2alpha.RequisitionSpecKt.eventGroupEntry
 import org.wfanet.measurement.api.v2alpha.RequisitionsGrpcKt.RequisitionsCoroutineImplBase
-import org.wfanet.measurement.api.v2alpha.RequisitionsGrpcKt.RequisitionsCoroutineStub
-import org.wfanet.measurement.api.v2alpha.UpdateEventGroupMetadataDescriptorRequest
 import org.wfanet.measurement.api.v2alpha.UpdateEventGroupRequest
-import org.wfanet.measurement.api.v2alpha.certificate
 import org.wfanet.measurement.api.v2alpha.copy
-import org.wfanet.measurement.api.v2alpha.dataProvider
+import org.wfanet.measurement.api.v2alpha.createEventGroupRequest
 import org.wfanet.measurement.api.v2alpha.differentialPrivacyParams
-import org.wfanet.measurement.api.v2alpha.elGamalPublicKey
 import org.wfanet.measurement.api.v2alpha.eventGroup
+import org.wfanet.measurement.api.v2alpha.eventGroupMetadata
 import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.SyntheticEventGroupSpec
-import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.SyntheticEventGroupSpecKt
-import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.TestMetadataMessage
-import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.copy
-import org.wfanet.measurement.api.v2alpha.event_templates.testing.Banner
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.Person
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.TestEvent
-import org.wfanet.measurement.api.v2alpha.event_templates.testing.Video
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.person
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.testEvent
-import org.wfanet.measurement.api.v2alpha.fulfillDirectRequisitionResponse
-import org.wfanet.measurement.api.v2alpha.getCertificateRequest
 import org.wfanet.measurement.api.v2alpha.liquidLegionsSketchParams
 import org.wfanet.measurement.api.v2alpha.listEventGroupsResponse
 import org.wfanet.measurement.api.v2alpha.listRequisitionsResponse
-import org.wfanet.measurement.api.v2alpha.measurementConsumer
 import org.wfanet.measurement.api.v2alpha.measurementSpec
-import org.wfanet.measurement.api.v2alpha.populationSpec
 import org.wfanet.measurement.api.v2alpha.protocolConfig
 import org.wfanet.measurement.api.v2alpha.refuseRequisitionRequest
 import org.wfanet.measurement.api.v2alpha.requisition
-import org.wfanet.measurement.api.v2alpha.requisitionSpec
 import org.wfanet.measurement.api.v2alpha.testing.MeasurementResultSubject.Companion.assertThat
 import org.wfanet.measurement.api.v2alpha.unpack
-import org.wfanet.measurement.common.HexString
-import org.wfanet.measurement.common.OpenEndTimeRange
-import org.wfanet.measurement.common.ProtoReflection
+import org.wfanet.measurement.api.v2alpha.updateEventGroupRequest
 import org.wfanet.measurement.common.crypto.Hashing
-import org.wfanet.measurement.common.crypto.SigningKeyHandle
-import org.wfanet.measurement.common.crypto.authorityKeyIdentifier
-import org.wfanet.measurement.common.crypto.readCertificateCollection
-import org.wfanet.measurement.common.crypto.subjectKeyIdentifier
-import org.wfanet.measurement.common.crypto.testing.loadSigningKey
-import org.wfanet.measurement.common.crypto.tink.TinkPrivateKeyHandle
-import org.wfanet.measurement.common.crypto.tink.loadPrivateKey
 import org.wfanet.measurement.common.crypto.tink.loadPublicKey
 import org.wfanet.measurement.common.flatten
-import org.wfanet.measurement.common.getRuntimePath
-import org.wfanet.measurement.common.grpc.testing.GrpcTestServerRule
-import org.wfanet.measurement.common.grpc.testing.mockService
-import org.wfanet.measurement.common.identity.externalIdToApiId
 import org.wfanet.measurement.common.pack
-import org.wfanet.measurement.common.readByteString
 import org.wfanet.measurement.common.testing.verifyAndCapture
 import org.wfanet.measurement.common.throttler.MinimumIntervalThrottler
-import org.wfanet.measurement.common.throttler.Throttler
-import org.wfanet.measurement.common.toProtoDate
 import org.wfanet.measurement.common.toProtoTime
 import org.wfanet.measurement.consent.client.common.toEncryptionPublicKey
 import org.wfanet.measurement.consent.client.duchy.computeRequisitionFingerprint
-import org.wfanet.measurement.consent.client.duchy.signElgamalPublicKey
-import org.wfanet.measurement.consent.client.measurementconsumer.decryptMetadata
 import org.wfanet.measurement.consent.client.measurementconsumer.decryptResult
 import org.wfanet.measurement.consent.client.measurementconsumer.encryptRequisitionSpec
 import org.wfanet.measurement.consent.client.measurementconsumer.signEncryptionPublicKey
 import org.wfanet.measurement.consent.client.measurementconsumer.signMeasurementSpec
 import org.wfanet.measurement.consent.client.measurementconsumer.signRequisitionSpec
-import org.wfanet.measurement.dataprovider.DataProviderData
 import org.wfanet.measurement.dataprovider.MeasurementResults
 import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.AcdpCharge
 import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.AgeGroup as PrivacyLandscapeAge
 import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.Gender as PrivacyLandscapeGender
-import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyBucketFilter
 import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyBucketGroup
-import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyBudgetManager
 import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.PrivacyLandscape.PRIVACY_BUCKET_VID_SAMPLE_WIDTH
-import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.testing.TestInMemoryBackingStore
-import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.testing.TestPrivacyBucketMapper
-import org.wfanet.measurement.eventdataprovider.requisition.v2alpha.common.InMemoryVidIndexMap
 import org.wfanet.measurement.integration.common.SyntheticGenerationSpecs
 import org.wfanet.measurement.loadtest.common.sampleVids
-import org.wfanet.measurement.loadtest.config.EventGroupMetadata
 import org.wfanet.measurement.loadtest.config.TestIdentifiers
 
-private const val MC_ID = "mc"
-private const val MC_NAME = "measurementConsumers/$MC_ID"
-private const val EDP_DISPLAY_NAME = "edp1"
-private val SECRET_FILES_PATH: Path =
-  checkNotNull(
-    getRuntimePath(
-      Paths.get("wfa_measurement_system", "src", "main", "k8s", "testing", "secretfiles")
-    )
-  )
-private const val EDP_ID = "someDataProvider"
-private const val EDP_NAME = "dataProviders/$EDP_ID"
-
-private const val LLV2_DECAY_RATE = 12.0
-private const val LLV2_MAX_SIZE = 100_000L
-private val NOISE_MECHANISM = ProtocolConfig.NoiseMechanism.DISCRETE_GAUSSIAN
-
-private val MEASUREMENT_CONSUMER_CERTIFICATE_DER =
-  SECRET_FILES_PATH.resolve("mc_cs_cert.der").toFile().readByteString()
-private const val MEASUREMENT_CONSUMER_NAME = "measurementConsumers/AAAAAAAAAHs"
-private const val MEASUREMENT_NAME = "$MC_NAME/measurements/BBBBBBBBBHs"
-private const val MEASUREMENT_CONSUMER_CERTIFICATE_NAME =
-  "$MEASUREMENT_CONSUMER_NAME/certificates/AAAAAAAAAcg"
-private val MEASUREMENT_CONSUMER_CERTIFICATE = certificate {
-  name = MEASUREMENT_CONSUMER_CERTIFICATE_NAME
-  x509Der = MEASUREMENT_CONSUMER_CERTIFICATE_DER
-}
-
-private val CONSENT_SIGNALING_ELGAMAL_PUBLIC_KEY = elGamalPublicKey {
-  ellipticCurveId = 415
-  generator = HexString("036B17D1F2E12C4247F8BCE6E563A440F277037D812DEB33A0F4A13945D898C296").bytes
-  element = HexString("0277BF406C5AA4376413E480E0AB8B0EFCA999D362204E6D1686E0BE567811604D").bytes
-}
-
-private val LAST_EVENT_DATE = LocalDate.now()
-private val FIRST_EVENT_DATE = LAST_EVENT_DATE.minusDays(1)
-private val TIME_RANGE = OpenEndTimeRange.fromClosedDateRange(FIRST_EVENT_DATE..LAST_EVENT_DATE)
-
-private const val DUCHY_ONE_ID = "worker1"
-private const val DUCHY_TWO_ID = "worker2"
 private const val RANDOM_SEED: Long = 0
 
 private const val RING_MODULUS = 127
@@ -233,119 +126,13 @@ private const val UNFULFILLABLE_EVENT_GROUP_ID = "unfulfillable"
 // Resource ID for EventGroup that fails Requisitions with DECLINED if used.
 private const val DECLINED_EVENT_GROUP_ID = "declined"
 
-@RunWith(JUnit4::class)
-class EdpSimulatorTest {
-  private val certificatesServiceMock: CertificatesCoroutineImplBase = mockService {
-    onBlocking {
-        getCertificate(eq(getCertificateRequest { name = MEASUREMENT_CONSUMER_CERTIFICATE_NAME }))
-      }
-      .thenReturn(MEASUREMENT_CONSUMER_CERTIFICATE)
-    onBlocking { getCertificate(eq(getCertificateRequest { name = DUCHY_ONE_CERTIFICATE.name })) }
-      .thenReturn(DUCHY_ONE_CERTIFICATE)
-    onBlocking { getCertificate(eq(getCertificateRequest { name = DUCHY_TWO_CERTIFICATE.name })) }
-      .thenReturn(DUCHY_TWO_CERTIFICATE)
-    onBlocking {
-        getCertificate(eq(getCertificateRequest { name = DATA_PROVIDER_CERTIFICATE.name }))
-      }
-      .thenReturn(DATA_PROVIDER_CERTIFICATE)
-    onBlocking {
-        getCertificate(eq(getCertificateRequest { name = DATA_PROVIDER_RESULT_CERTIFICATE.name }))
-      }
-      .thenReturn(DATA_PROVIDER_RESULT_CERTIFICATE)
-  }
-  private val dataProvidersServiceMock: DataProvidersCoroutineImplBase = mockService {
-    onBlocking { replaceDataAvailabilityInterval(any()) }
-      .thenAnswer {
-        val request = it.arguments[0] as ReplaceDataAvailabilityIntervalRequest
-        dataProvider { dataAvailabilityInterval = request.dataAvailabilityInterval }
-      }
-  }
-  private val measurementConsumersServiceMock:
-    MeasurementConsumersGrpcKt.MeasurementConsumersCoroutineImplBase =
-    mockService {
-      onBlocking { getMeasurementConsumer(any()) }
-        .thenReturn(
-          measurementConsumer {
-            publicKey = signEncryptionPublicKey(MC_PUBLIC_KEY, MC_SIGNING_KEY)
-            certificate = MEASUREMENT_CONSUMER_CERTIFICATE_NAME
-          }
-        )
-    }
-  private val eventGroupsServiceMock: EventGroupsCoroutineImplBase = mockService {
-    onBlocking { getEventGroup(any()) }
-      .thenAnswer { invocation ->
-        val request = invocation.getArgument<GetEventGroupRequest>(0)
-        eventGroup {
-          name = request.name
-          eventGroupReferenceId = TestIdentifiers.SIMULATOR_EVENT_GROUP_REFERENCE_ID_PREFIX
-        }
-      }
-  }
-  private val eventGroupMetadataDescriptorsServiceMock:
-    EventGroupMetadataDescriptorsCoroutineImplBase =
-    mockService {
-      onBlocking { createEventGroupMetadataDescriptor(any()) }
-        .thenAnswer { invocation ->
-          val request = invocation.getArgument<CreateEventGroupMetadataDescriptorRequest>(0)
-          request.eventGroupMetadataDescriptor.copy { name = EVENT_GROUP_METADATA_DESCRIPTOR_NAME }
-        }
-    }
-  private val requisitionsServiceMock: RequisitionsCoroutineImplBase = mockService {
-    onBlocking { listRequisitions(any()) }
-      .thenReturn(listRequisitionsResponse { requisitions += REQUISITION })
-    onBlocking { fulfillDirectRequisition(any()) }.thenReturn(fulfillDirectRequisitionResponse {})
-  }
-
-  private val fakeRequisitionFulfillmentService = FakeRequisitionFulfillmentService()
-
-  @get:Rule
-  val grpcTestServerRule = GrpcTestServerRule {
-    addService(measurementConsumersServiceMock)
-    addService(certificatesServiceMock)
-    addService(dataProvidersServiceMock)
-    addService(eventGroupsServiceMock)
-    addService(eventGroupMetadataDescriptorsServiceMock)
-    addService(requisitionsServiceMock)
-    addService(fakeRequisitionFulfillmentService)
-  }
-
-  private val measurementConsumersStub by lazy {
-    MeasurementConsumersCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val certificatesStub: CertificatesCoroutineStub by lazy {
-    CertificatesCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val dataProvidersStub: DataProvidersCoroutineStub by lazy {
-    DataProvidersCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val eventGroupsStub: EventGroupsCoroutineStub by lazy {
-    EventGroupsCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val eventGroupMetadataDescriptorsStub by lazy {
-    EventGroupMetadataDescriptorsCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val requisitionsStub: RequisitionsCoroutineStub by lazy {
-    RequisitionsCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val requisitionFulfillmentStub: RequisitionFulfillmentCoroutineStub by lazy {
-    RequisitionFulfillmentCoroutineStub(grpcTestServerRule.channel)
-  }
-
-  private val requisitionFulfillmentStubMap =
-    mapOf(
-      DuchyKey(DUCHY_ONE_ID).toName() to requisitionFulfillmentStub,
-      DuchyKey(DUCHY_TWO_ID).toName() to requisitionFulfillmentStub,
-    )
-
-  private val backingStore = TestInMemoryBackingStore()
-  private val privacyBudgetManager =
-    PrivacyBudgetManager(PrivacyBucketFilter(TestPrivacyBucketMapper()), backingStore, 10.0f, 0.02f)
+class EdpSimulatorTest : AbstractEdpSimulatorTest() {
+  private data class EventGroupOptions(
+    override val referenceIdSuffix: String,
+    override val syntheticDataSpec: SyntheticEventGroupSpec,
+    override val mediaTypes: Set<MediaType>,
+    override val metadata: EventGroupMetadata?,
+  ) : EdpSimulator.EventGroupOptions
 
   private fun generateEvents(
     vidRange: LongRange,
@@ -364,79 +151,84 @@ class EdpSimulatorTest {
   }
 
   @Test
-  fun `ensureEventGroup creates EventGroup and EventGroupMetadataDescriptor`() {
-    val knownEventGroupMetadataTypes = listOf(SyntheticEventGroupSpec.getDescriptor().file)
+  fun `ensureEventGroups creates EventGroup`() {
+    val referenceIdSuffix = "-001"
     val edpSimulator =
       EdpSimulator(
         EDP_DATA,
         MEASUREMENT_CONSUMER_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(
+          EventGroupOptions(
+            referenceIdSuffix,
+            SYNTHETIC_DATA_SPEC,
+            MEDIA_TYPES,
+            EVENT_GROUP_METADATA,
+          )
+        ),
         InMemoryEventQuery(emptyList()),
         dummyThrottler,
         privacyBudgetManager,
         TRUSTED_CERTIFICATES,
         HMSS_VID_INDEX_MAP,
-        knownEventGroupMetadataTypes = knownEventGroupMetadataTypes,
       )
 
-    runBlocking { edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, SYNTHETIC_DATA_SPEC) }
+    runBlocking { edpSimulator.ensureEventGroups() }
 
-    // Verify metadata descriptor is created.
-    verifyBlocking(eventGroupMetadataDescriptorsServiceMock) {
-      createEventGroupMetadataDescriptor(any())
-    }
-
-    // Verify EventGroup metadata.
-    val createRequest: CreateEventGroupRequest =
+    val request: CreateEventGroupRequest =
       verifyAndCapture(eventGroupsServiceMock, EventGroupsCoroutineImplBase::createEventGroup)
-    val metadata: EventGroup.Metadata =
-      decryptMetadata(createRequest.eventGroup.encryptedMetadata, MC_PRIVATE_KEY)
-    assertThat(metadata.eventGroupMetadataDescriptor)
-      .isEqualTo(EVENT_GROUP_METADATA_DESCRIPTOR_NAME)
-    assertThat(metadata.metadata.unpack(SyntheticEventGroupSpec::class.java))
-      .isEqualTo(SYNTHETIC_DATA_SPEC)
-
-    // Verify EventGroup has correct template types.
-    assertThat(createRequest.eventGroup.eventTemplatesList.map { it.type })
-      .containsAtLeast(
-        Person.getDescriptor().fullName,
-        Video.getDescriptor().fullName,
-        Banner.getDescriptor().fullName,
+    assertThat(request)
+      .isEqualTo(
+        createEventGroupRequest {
+          parent = EDP_NAME
+          eventGroup = eventGroup {
+            measurementConsumer = MEASUREMENT_CONSUMER_NAME
+            eventGroupReferenceId =
+              "${TestIdentifiers.SIMULATOR_EVENT_GROUP_REFERENCE_ID_PREFIX}-${EDP_DATA.displayName}${referenceIdSuffix}"
+            mediaTypes += MEDIA_TYPES
+            eventGroupMetadata = EVENT_GROUP_METADATA
+            dataAvailabilityInterval = DATA_AVAILABILITY_INTERVAL
+          }
+        }
       )
   }
 
   @Test
   fun `ensureEventGroup updates EventGroup`() {
+    val referenceIdSuffix = "-001"
+    val eventGroup = eventGroup {
+      name = EVENT_GROUP_NAME
+      measurementConsumer = MEASUREMENT_CONSUMER_NAME
+      eventGroupReferenceId =
+        "${TestIdentifiers.SIMULATOR_EVENT_GROUP_REFERENCE_ID_PREFIX}-${EDP_DATA.displayName}${referenceIdSuffix}"
+    }
     eventGroupsServiceMock.stub {
       onBlocking { listEventGroups(any()) }
-        .thenReturn(
-          listEventGroupsResponse {
-            eventGroups += eventGroup {
-              name = EVENT_GROUP_NAME
-              measurementConsumer = MEASUREMENT_CONSUMER_NAME
-              eventGroupReferenceId =
-                "${TestIdentifiers.SIMULATOR_EVENT_GROUP_REFERENCE_ID_PREFIX}-${EDP_DATA.displayName}"
-            }
-          }
-        )
+        .thenReturn(listEventGroupsResponse { eventGroups += eventGroup })
     }
     val edpSimulator =
       EdpSimulator(
         EDP_DATA,
         MEASUREMENT_CONSUMER_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(
+          EventGroupOptions(
+            referenceIdSuffix,
+            SYNTHETIC_DATA_SPEC,
+            MEDIA_TYPES,
+            EVENT_GROUP_METADATA,
+          )
+        ),
         InMemoryEventQuery(emptyList()),
         dummyThrottler,
         privacyBudgetManager,
@@ -444,160 +236,22 @@ class EdpSimulatorTest {
         HMSS_VID_INDEX_MAP,
       )
 
-    runBlocking { edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, SYNTHETIC_DATA_SPEC) }
+    runBlocking { edpSimulator.ensureEventGroups() }
 
     // Verify EventGroup metadata has correct type.
     val updateRequest: UpdateEventGroupRequest =
       verifyAndCapture(eventGroupsServiceMock, EventGroupsCoroutineImplBase::updateEventGroup)
-    val metadata: EventGroup.Metadata =
-      decryptMetadata(updateRequest.eventGroup.encryptedMetadata, MC_PRIVATE_KEY)
-    assertThat(metadata.eventGroupMetadataDescriptor)
-      .isEqualTo(EVENT_GROUP_METADATA_DESCRIPTOR_NAME)
-    assertThat(metadata.metadata.unpack(SyntheticEventGroupSpec::class.java))
-      .isEqualTo(SYNTHETIC_DATA_SPEC)
-
-    // Verify EventGroup has correct template types.
-    assertThat(updateRequest.eventGroup.eventTemplatesList.map { it.type })
-      .containsAtLeast(
-        Person.getDescriptor().fullName,
-        Video.getDescriptor().fullName,
-        Banner.getDescriptor().fullName,
-      )
-  }
-
-  @Test
-  fun `ensureEventGroup updates EventGroupMetadataDescriptor`() {
-    eventGroupMetadataDescriptorsServiceMock.stub {
-      onBlocking { createEventGroupMetadataDescriptor(any()) }
-        .thenAnswer { invocation ->
-          val request = invocation.getArgument<CreateEventGroupMetadataDescriptorRequest>(0)
-          request.eventGroupMetadataDescriptor.copy {
-            descriptorSet =
-              ProtoReflection.buildFileDescriptorSet(TestMetadataMessage.getDescriptor())
-          }
+    assertThat(updateRequest)
+      .isEqualTo(
+        updateEventGroupRequest {
+          this.eventGroup =
+            eventGroup.copy {
+              mediaTypes += MEDIA_TYPES
+              eventGroupMetadata = EVENT_GROUP_METADATA
+              dataAvailabilityInterval = DATA_AVAILABILITY_INTERVAL
+            }
         }
-    }
-    val edpSimulator =
-      EdpSimulator(
-        EDP_DATA,
-        MEASUREMENT_CONSUMER_NAME,
-        measurementConsumersStub,
-        certificatesStub,
-        dataProvidersStub,
-        eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
-        requisitionsStub,
-        requisitionFulfillmentStubMap,
-        InMemoryEventQuery(emptyList()),
-        dummyThrottler,
-        privacyBudgetManager,
-        TRUSTED_CERTIFICATES,
-        HMSS_VID_INDEX_MAP,
       )
-
-    runBlocking { edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, SYNTHETIC_DATA_SPEC) }
-
-    val updateRequest: UpdateEventGroupMetadataDescriptorRequest =
-      verifyAndCapture(
-        eventGroupMetadataDescriptorsServiceMock,
-        EventGroupMetadataDescriptorsCoroutineImplBase::updateEventGroupMetadataDescriptor,
-      )
-    assertThat(updateRequest.eventGroupMetadataDescriptor.descriptorSet)
-      .isEqualTo(ProtoReflection.buildFileDescriptorSet(SyntheticEventGroupSpec.getDescriptor()))
-  }
-
-  @Test
-  fun `ensureEventGroups creates multiple EventGroups`() {
-    val edpSimulator =
-      EdpSimulator(
-        EDP_DATA,
-        MEASUREMENT_CONSUMER_NAME,
-        measurementConsumersStub,
-        certificatesStub,
-        dataProvidersStub,
-        eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
-        requisitionsStub,
-        requisitionFulfillmentStubMap,
-        InMemoryEventQuery(emptyList()),
-        dummyThrottler,
-        privacyBudgetManager,
-        TRUSTED_CERTIFICATES,
-        HMSS_VID_INDEX_MAP,
-      )
-    val metadataByReferenceIdSuffix =
-      mapOf(
-        "-foo" to SyntheticGenerationSpecs.SYNTHETIC_DATA_SPECS_SMALL[0],
-        "-bar" to SyntheticGenerationSpecs.SYNTHETIC_DATA_SPECS_SMALL[1],
-      )
-
-    runBlocking {
-      edpSimulator.ensureEventGroups(TEST_EVENT_TEMPLATES, metadataByReferenceIdSuffix)
-    }
-
-    // Verify metadata descriptor set contains synthetic data spec.
-    val createDescriptorRequest: CreateEventGroupMetadataDescriptorRequest =
-      verifyAndCapture(
-        eventGroupMetadataDescriptorsServiceMock,
-        EventGroupMetadataDescriptorsCoroutineImplBase::createEventGroupMetadataDescriptor,
-      )
-    val descriptors =
-      ProtoReflection.buildDescriptors(
-        listOf(createDescriptorRequest.eventGroupMetadataDescriptor.descriptorSet)
-      )
-    assertThat(descriptors.map { it.fullName })
-      .contains(SyntheticEventGroupSpec.getDescriptor().fullName)
-
-    // Verify EventGroup metadata.
-    val createRequests: List<CreateEventGroupRequest> =
-      verifyAndCapture(
-        eventGroupsServiceMock,
-        EventGroupsCoroutineImplBase::createEventGroup,
-        times(2),
-      )
-    for (createRequest in createRequests) {
-      val metadata: EventGroup.Metadata =
-        decryptMetadata(createRequest.eventGroup.encryptedMetadata, MC_PRIVATE_KEY)
-      assertThat(metadata.eventGroupMetadataDescriptor)
-        .isEqualTo(EVENT_GROUP_METADATA_DESCRIPTOR_NAME)
-      assertThat(metadata.metadata.unpack(SyntheticEventGroupSpec::class.java))
-        .isEqualTo(
-          metadataByReferenceIdSuffix.getValue(
-            EdpSimulator.getEventGroupReferenceIdSuffix(createRequest.eventGroup, EDP_DISPLAY_NAME)
-          )
-        )
-    }
-  }
-
-  @Test
-  fun `ensureEventGroups throws IllegalArgumentException when metadata message types mismatch`() {
-    val edpSimulator =
-      EdpSimulator(
-        EDP_DATA,
-        MEASUREMENT_CONSUMER_NAME,
-        measurementConsumersStub,
-        certificatesStub,
-        dataProvidersStub,
-        eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
-        requisitionsStub,
-        requisitionFulfillmentStubMap,
-        InMemoryEventQuery(emptyList()),
-        dummyThrottler,
-        privacyBudgetManager,
-        TRUSTED_CERTIFICATES,
-        HMSS_VID_INDEX_MAP,
-      )
-    val metadataByReferenceIdSuffix = mapOf("-foo" to SYNTHETIC_DATA_SPEC, "-bar" to TEST_METADATA)
-
-    val exception =
-      assertFailsWith<IllegalArgumentException> {
-        runBlocking {
-          edpSimulator.ensureEventGroups(TEST_EVENT_TEMPLATES, metadataByReferenceIdSuffix)
-        }
-      }
-
-    assertThat(exception).hasMessageThat().contains("type")
   }
 
   @Test
@@ -638,13 +292,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         "measurementConsumers/differentMcId",
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         InMemoryEventQuery(allEvents),
         MinimumIntervalThrottler(Clock.systemUTC(), Duration.ofMillis(1000)),
         privacyBudgetManager,
@@ -653,7 +307,7 @@ class EdpSimulatorTest {
       )
 
     runBlocking {
-      edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, SYNTHETIC_DATA_SPEC)
+      edpSimulator.ensureEventGroups()
       edpSimulator.executeRequisitionFulfillingWorkflow()
     }
 
@@ -677,13 +331,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         InMemoryEventQuery(emptyList()),
         dummyThrottler,
         privacyBudgetManager,
@@ -731,13 +385,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         InMemoryEventQuery(emptyList()),
         dummyThrottler,
         privacyBudgetManager,
@@ -810,13 +464,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         InMemoryEventQuery(allEvents),
         dummyThrottler,
         privacyBudgetManager,
@@ -865,13 +519,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         InMemoryEventQuery(emptyList()),
         dummyThrottler,
         privacyBudgetManager,
@@ -957,13 +611,13 @@ class EdpSimulatorTest {
         EdpSimulator(
           EDP_DATA,
           MC_NAME,
-          measurementConsumersStub,
           certificatesStub,
           dataProvidersStub,
           eventGroupsStub,
-          eventGroupMetadataDescriptorsStub,
           requisitionsStub,
           requisitionFulfillmentStubMap,
+          SYNTHETIC_DATA_TIME_ZONE,
+          listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
           InMemoryEventQuery(allEvents),
           dummyThrottler,
           privacyBudgetManager,
@@ -971,7 +625,7 @@ class EdpSimulatorTest {
           HMSS_VID_INDEX_MAP,
         )
       runBlocking {
-        edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, TEST_METADATA)
+        edpSimulator.ensureEventGroups()
         edpSimulator.executeRequisitionFulfillingWorkflow()
       }
 
@@ -1086,13 +740,13 @@ class EdpSimulatorTest {
         EdpSimulator(
           EDP_DATA,
           MC_NAME,
-          measurementConsumersStub,
           certificatesStub,
           dataProvidersStub,
           eventGroupsStub,
-          eventGroupMetadataDescriptorsStub,
           requisitionsStub,
           requisitionFulfillmentStubMap,
+          SYNTHETIC_DATA_TIME_ZONE,
+          listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
           InMemoryEventQuery(allEvents),
           dummyThrottler,
           privacyBudgetManager,
@@ -1100,7 +754,7 @@ class EdpSimulatorTest {
           HMSS_VID_INDEX_MAP,
         )
       runBlocking {
-        edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, TEST_METADATA)
+        edpSimulator.ensureEventGroups()
         edpSimulator.executeRequisitionFulfillingWorkflow()
       }
 
@@ -1235,13 +889,13 @@ class EdpSimulatorTest {
         EdpSimulator(
           EDP_DATA,
           MC_NAME,
-          measurementConsumersStub,
           certificatesStub,
           dataProvidersStub,
           eventGroupsStub,
-          eventGroupMetadataDescriptorsStub,
           requisitionsStub,
           requisitionFulfillmentStubMap,
+          SYNTHETIC_DATA_TIME_ZONE,
+          listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
           InMemoryEventQuery(allEvents),
           dummyThrottler,
           privacyBudgetManager,
@@ -1249,7 +903,7 @@ class EdpSimulatorTest {
           HMSS_VID_INDEX_MAP,
         )
       runBlocking {
-        edpSimulator.ensureEventGroup(TEST_EVENT_TEMPLATES, TEST_METADATA)
+        edpSimulator.ensureEventGroups()
         edpSimulator.executeRequisitionFulfillingWorkflow()
       }
 
@@ -1342,13 +996,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -1400,13 +1054,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1460,13 +1114,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1531,13 +1185,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1599,13 +1253,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1667,13 +1321,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1733,13 +1387,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1799,13 +1453,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1865,13 +1519,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1931,13 +1585,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         eventQueryMock,
         dummyThrottler,
         privacyBudgetManager,
@@ -1994,13 +1648,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2051,13 +1705,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2155,13 +1809,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2226,13 +1880,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2297,13 +1951,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2355,13 +2009,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2420,13 +2074,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2485,13 +2139,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2551,13 +2205,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2614,13 +2268,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2678,13 +2332,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2745,13 +2399,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2807,13 +2461,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2866,13 +2520,13 @@ class EdpSimulatorTest {
       EdpSimulator(
         EDP_DATA,
         MC_NAME,
-        measurementConsumersStub,
         certificatesStub,
         dataProvidersStub,
         eventGroupsStub,
-        eventGroupMetadataDescriptorsStub,
         requisitionsStub,
         requisitionFulfillmentStubMap,
+        SYNTHETIC_DATA_TIME_ZONE,
+        listOf(EventGroupOptions("", SYNTHETIC_DATA_SPEC, MEDIA_TYPES, EVENT_GROUP_METADATA)),
         syntheticGeneratorEventQuery,
         dummyThrottler,
         privacyBudgetManager,
@@ -2902,134 +2556,27 @@ class EdpSimulatorTest {
     verifyBlocking(requisitionsServiceMock, never()) { fulfillDirectRequisition(any()) }
   }
 
-  private class FakeRequisitionFulfillmentService : RequisitionFulfillmentCoroutineImplBase() {
-    data class FulfillRequisitionInvocation(val requests: List<FulfillRequisitionRequest>)
-
-    private val _fullfillRequisitionInvocations = mutableListOf<FulfillRequisitionInvocation>()
-    val fullfillRequisitionInvocations: List<FulfillRequisitionInvocation>
-      get() = _fullfillRequisitionInvocations
-
-    override suspend fun fulfillRequisition(
-      requests: Flow<FulfillRequisitionRequest>
-    ): FulfillRequisitionResponse {
-      // Consume flow before returning.
-      _fullfillRequisitionInvocations.add(FulfillRequisitionInvocation(requests.toList()))
-      return FulfillRequisitionResponse.getDefaultInstance()
-    }
-  }
-
   companion object {
-    private const val EVENT_GROUP_METADATA_DESCRIPTOR_NAME =
-      "dataProviders/foo/eventGroupMetadataDescriptors/bar"
-
-    private val MC_SIGNING_KEY = loadSigningKey("${MC_ID}_cs_cert.der", "${MC_ID}_cs_private.der")
-    private val DUCHY_ONE_SIGNING_KEY =
-      loadSigningKey("${DUCHY_ONE_ID}_cs_cert.der", "${DUCHY_ONE_ID}_cs_private.der")
-    private val DUCHY_TWO_SIGNING_KEY =
-      loadSigningKey("${DUCHY_TWO_ID}_cs_cert.der", "${DUCHY_TWO_ID}_cs_private.der")
-
-    private val DUCHY_ONE_NAME = DuchyKey(DUCHY_ONE_ID).toName()
-    private val DUCHY_TWO_NAME = DuchyKey(DUCHY_TWO_ID).toName()
-    private val DUCHY_ONE_CERTIFICATE = certificate {
-      name = DuchyCertificateKey(DUCHY_ONE_ID, externalIdToApiId(6L)).toName()
-      x509Der = DUCHY_ONE_SIGNING_KEY.certificate.encoded.toByteString()
+    private val MEDIA_TYPES = setOf(MediaType.VIDEO, MediaType.DISPLAY)
+    private val DATA_AVAILABILITY_INTERVAL = interval {
+      startTime = FIRST_EVENT_DATE.atStartOfDay(SYNTHETIC_DATA_TIME_ZONE).toInstant().toProtoTime()
+      endTime =
+        LAST_EVENT_DATE.plusDays(1).atStartOfDay(SYNTHETIC_DATA_TIME_ZONE).toInstant().toProtoTime()
     }
-    private val DUCHY_TWO_CERTIFICATE = certificate {
-      name = DuchyCertificateKey(DUCHY_TWO_ID, externalIdToApiId(6L)).toName()
-      x509Der = DUCHY_TWO_SIGNING_KEY.certificate.encoded.toByteString()
+    private val EVENT_GROUP_METADATA = eventGroupMetadata {
+      adMetadata =
+        EventGroupMetadataKt.adMetadata {
+          campaignMetadata =
+            EventGroupMetadataKt.AdMetadataKt.campaignMetadata {
+              brandName = "Log by Blammo!"
+              campaignName = "Better Than Bad"
+            }
+        }
     }
 
     private val DUCHY1_ENCRYPTION_PUBLIC_KEY =
       loadPublicKey(SECRET_FILES_PATH.resolve("mc_enc_public.tink").toFile())
         .toEncryptionPublicKey()
-
-    private val EDP_SIGNING_KEY =
-      loadSigningKey("${EDP_DISPLAY_NAME}_cs_cert.der", "${EDP_DISPLAY_NAME}_cs_private.der")
-    private val EDP_RESULT_SIGNING_KEY =
-      loadSigningKey(
-        "${EDP_DISPLAY_NAME}_result_cs_cert.der",
-        "${EDP_DISPLAY_NAME}_result_cs_private.der",
-      )
-    private val DATA_PROVIDER_CERTIFICATE_KEY =
-      DataProviderCertificateKey(EDP_ID, externalIdToApiId(8L))
-    private val DATA_PROVIDER_RESULT_CERTIFICATE_KEY =
-      DataProviderCertificateKey(EDP_ID, externalIdToApiId(9L))
-
-    private val DATA_PROVIDER_CERTIFICATE = certificate {
-      name = DATA_PROVIDER_CERTIFICATE_KEY.toName()
-      x509Der = EDP_SIGNING_KEY.certificate.encoded.toByteString()
-      subjectKeyIdentifier = EDP_SIGNING_KEY.certificate.subjectKeyIdentifier!!
-    }
-    private val DATA_PROVIDER_RESULT_CERTIFICATE = certificate {
-      name = DATA_PROVIDER_RESULT_CERTIFICATE_KEY.toName()
-      x509Der = EDP_RESULT_SIGNING_KEY.certificate.encoded.toByteString()
-      subjectKeyIdentifier = EDP_RESULT_SIGNING_KEY.certificate.subjectKeyIdentifier!!
-    }
-    private val EDP_DATA =
-      DataProviderData(
-        EDP_NAME,
-        EDP_DISPLAY_NAME,
-        loadEncryptionPrivateKey("${EDP_DISPLAY_NAME}_enc_private.tink"),
-        EDP_RESULT_SIGNING_KEY,
-        DATA_PROVIDER_RESULT_CERTIFICATE_KEY,
-      )
-
-    private val MC_PUBLIC_KEY =
-      loadPublicKey(SECRET_FILES_PATH.resolve("mc_enc_public.tink").toFile())
-        .toEncryptionPublicKey()
-    private val MC_PRIVATE_KEY =
-      loadPrivateKey(SECRET_FILES_PATH.resolve("mc_enc_private.tink").toFile())
-    private val DATA_PROVIDER_PUBLIC_KEY =
-      loadPublicKey(SECRET_FILES_PATH.resolve("${EDP_DISPLAY_NAME}_enc_public.tink").toFile())
-        .toEncryptionPublicKey()
-
-    private const val EVENT_GROUP_NAME = "$EDP_NAME/eventGroups/name"
-    private val REQUISITION_SPEC = requisitionSpec {
-      events =
-        RequisitionSpecKt.events {
-          eventGroups += eventGroupEntry {
-            key = EVENT_GROUP_NAME
-            value =
-              RequisitionSpecKt.EventGroupEntryKt.value {
-                collectionInterval = interval {
-                  startTime = TIME_RANGE.start.toProtoTime()
-                  endTime = TIME_RANGE.endExclusive.toProtoTime()
-                }
-                filter = eventFilter {
-                  expression =
-                    "person.age_group == ${Person.AgeGroup.YEARS_18_TO_34_VALUE} && " +
-                      "person.gender == ${Person.Gender.FEMALE_VALUE}"
-                }
-              }
-          }
-        }
-      measurementPublicKey = MC_PUBLIC_KEY.pack()
-      nonce = Random.Default.nextLong()
-    }
-
-    private val ENCRYPTED_REQUISITION_SPEC =
-      encryptRequisitionSpec(
-        signRequisitionSpec(REQUISITION_SPEC, MC_SIGNING_KEY),
-        DATA_PROVIDER_PUBLIC_KEY,
-      )
-
-    private val OUTPUT_DP_PARAMS = differentialPrivacyParams {
-      epsilon = 1.0
-      delta = 1E-12
-    }
-    private val MEASUREMENT_SPEC = measurementSpec {
-      measurementPublicKey = MC_PUBLIC_KEY.pack()
-      reachAndFrequency = reachAndFrequency {
-        reachPrivacyParams = OUTPUT_DP_PARAMS
-        frequencyPrivacyParams = OUTPUT_DP_PARAMS
-        maximumFrequency = 10
-      }
-      vidSamplingInterval = vidSamplingInterval {
-        start = 0.0f
-        width = 1.0f
-      }
-      nonceHashes += Hashing.hashSha256(REQUISITION_SPEC.nonce)
-    }
 
     private val REACH_ONLY_MEASUREMENT_SPEC =
       MEASUREMENT_SPEC.copy {
@@ -3050,44 +2597,6 @@ class EdpSimulatorTest {
         width = 1.0f
       }
       nonceHashes += Hashing.hashSha256(REQUISITION_SPEC.nonce)
-    }
-
-    private val LIQUID_LEGIONS_SKETCH_PARAMS = liquidLegionsSketchParams {
-      decayRate = LLV2_DECAY_RATE
-      maxSize = LLV2_MAX_SIZE
-      samplingIndicatorSize = 10_000_000
-    }
-
-    private val REQUISITION = requisition {
-      name = "${EDP_NAME}/requisitions/foo"
-      measurement = MEASUREMENT_NAME
-      state = Requisition.State.UNFULFILLED
-      measurementConsumerCertificate = MEASUREMENT_CONSUMER_CERTIFICATE_NAME
-      measurementSpec = signMeasurementSpec(MEASUREMENT_SPEC, MC_SIGNING_KEY)
-      encryptedRequisitionSpec = ENCRYPTED_REQUISITION_SPEC
-      protocolConfig = protocolConfig {
-        protocols +=
-          ProtocolConfigKt.protocol {
-            liquidLegionsV2 =
-              ProtocolConfigKt.liquidLegionsV2 {
-                noiseMechanism = NOISE_MECHANISM
-                sketchParams = LIQUID_LEGIONS_SKETCH_PARAMS
-                ellipticCurveId = 415
-              }
-          }
-      }
-      dataProviderCertificate = DATA_PROVIDER_CERTIFICATE.name
-      dataProviderPublicKey = DATA_PROVIDER_PUBLIC_KEY.pack()
-      duchies += duchyEntry {
-        key = DUCHY_ONE_NAME
-        value = value {
-          duchyCertificate = DUCHY_ONE_CERTIFICATE.name
-          liquidLegionsV2 = liquidLegionsV2 {
-            elGamalPublicKey =
-              signElgamalPublicKey(CONSENT_SIGNALING_ELGAMAL_PUBLIC_KEY, DUCHY_ONE_SIGNING_KEY)
-          }
-        }
-      }
     }
 
     private val DUCHY_ENTRY_ONE = duchyEntry {
@@ -3129,52 +2638,15 @@ class EdpSimulatorTest {
       duchies += DUCHY_ENTRY_TWO
     }
 
-    private val TRUSTED_CERTIFICATES: Map<ByteString, X509Certificate> =
-      readCertificateCollection(SECRET_FILES_PATH.resolve("edp_trusted_certs.pem").toFile())
-        .associateBy { requireNotNull(it.authorityKeyIdentifier) }
-
-    private val TEST_EVENT_TEMPLATES = EdpSimulator.buildEventTemplates(TestEvent.getDescriptor())
-    private val TEST_METADATA = EventGroupMetadata.testMetadata(1)
-
-    private val SYNTHETIC_DATA_SPEC =
-      SyntheticGenerationSpecs.SYNTHETIC_DATA_SPECS_SMALL.first().copy {
-        dateSpecs.forEachIndexed { index, dateSpec ->
-          dateSpecs[index] =
-            dateSpec.copy {
-              dateRange =
-                SyntheticEventGroupSpecKt.DateSpecKt.dateRange {
-                  start = FIRST_EVENT_DATE.toProtoDate()
-                  endExclusive = (LAST_EVENT_DATE.plusDays(1)).toProtoDate()
-                }
-            }
-        }
-      }
     private val syntheticGeneratorEventQuery =
       object :
         SyntheticGeneratorEventQuery(
           SyntheticGenerationSpecs.SYNTHETIC_POPULATION_SPEC_SMALL,
           TestEvent.getDescriptor(),
+          SYNTHETIC_DATA_TIME_ZONE,
         ) {
         override fun getSyntheticDataSpec(eventGroup: EventGroup): SyntheticEventGroupSpec {
           return SYNTHETIC_DATA_SPEC
-        }
-      }
-
-    private val POPULATION_SPEC = populationSpec {
-      subpopulations += subPopulation {
-        vidRanges += vidRange {
-          startVid = 1L
-          endVidInclusive = 1000L
-        }
-      }
-    }
-    private val HMSS_VID_INDEX_MAP = InMemoryVidIndexMap.build(POPULATION_SPEC)
-
-    /** Dummy [Throttler] for satisfying signatures without being used. */
-    private val dummyThrottler =
-      object : Throttler {
-        override suspend fun <T> onReady(block: suspend () -> T): T {
-          throw UnsupportedOperationException("Should not be called")
         }
       }
 
@@ -3199,23 +2671,9 @@ class EdpSimulatorTest {
         }
       }
 
-    private fun loadSigningKey(
-      certDerFileName: String,
-      privateKeyDerFileName: String,
-    ): SigningKeyHandle {
-      return loadSigningKey(
-        SECRET_FILES_PATH.resolve(certDerFileName).toFile(),
-        SECRET_FILES_PATH.resolve(privateKeyDerFileName).toFile(),
-      )
-    }
-
     private const val REACH_TOLERANCE = 1.0
     private const val FREQUENCY_DISTRIBUTION_TOLERANCE = 1.0
     private const val IMPRESSION_TOLERANCE = 1.0
-
-    private fun loadEncryptionPrivateKey(fileName: String): TinkPrivateKeyHandle {
-      return loadPrivateKey(SECRET_FILES_PATH.resolve(fileName).toFile())
-    }
 
     private fun computeExpectedReach(
       eventGroupsList: List<RequisitionSpec.EventGroupEntry>,
